Updates to naming convention, added SettingsMenu component into each page (which includes the theme switcher and logout button)
